Remove excess code
