+++ test/featuretests/conf/router.yaml
@@ -1,0 +1,44 @@
+host: 'regress_router'
+router_port: '6432'
+admin_console_port: '7432'
+grpc_api_port: '7000'
+router_mode: PROXY
+log_level: fatal
+time_quantiles:
+  - 0.75
+world_shard_fallback: true
+show_notice_messages: true
+frontend_rules:
+  - db: regress
+    usr: regress
+    pool_mode: TRANSACTION
+    auth_rule:
+      auth_method: ok
+shards:
+  sh1:
+    db: regress
+    usr: regress
+    pwd: 12345678
+    type: DATA
+    hosts:
+      - 'spqr_shard_1:6432'
+  sh2:
+    db: regress
+    usr: regress
+    pwd: 12345678
+    type: DATA
+    hosts:
+      - 'spqr_shard_2:6432'
+
+backend_rules:
+  - db: regress
+    usr: regress
+    pool_discard: true
+    pool_rollback: true
+    auth_rules:
+      sh1:
+        auth_method: md5
+        password: 12345678
+      sh2:
+        auth_method: md5
+        password: 12345678
