+++ test/featuretests/docker-compose.yaml
@@ -34,11 +34,11 @@ services:
   router:
     build:
       dockerfile: ./docker/router/Dockerfile
-      context: ../../../
+      context: ../../
     ports:
       - "6432:6432"
     environment:
-      - SPQR_CONFIG=/spqr/test/regress/conf/router.yaml
+      - SPQR_CONFIG=/spqr/test/featuretests/conf/router.yaml
     hostname: regress_router
     container_name: regress_router
     depends_on:
@@ -50,12 +50,12 @@ services:
   coordinator:
     build:
       dockerfile: ./docker/coordinator/Dockerfile
-      context: ../../../
+      context: ../../
     ports:
       - "7002:7002"
       - "7003:7003"
     environment:
-      - SPQR_CONFIG=/spqr/test/regress/conf/coordinator.yaml
+      - SPQR_CONFIG=/spqr/test/featuretests/conf/coordinator.yaml
     hostname: regress_coordinator
     container_name: regress_coordinator
     depends_on:
