+++ test/featuretests/features/base.feature
@@ -13,11 +13,8 @@ Feature: Basic test
     """
     When I run SQL on host "router"
     """
-    SELECT 1
     """
     Then command return code should be "0"
     And SQL result should match regexp
     """
-    1
-    1
     """
