--- test/featuretests/images/base/supervisor.conf
@@ -1,20 +1,0 @@
-[unix_http_server]
-file=/var/run/supervisor.sock
-chmod=0777
-
-[supervisord]
-logfile=/var/log/supervisor.log
-logfile_maxbytes=0
-log_level=debug
-pidfile=/var/run/supervisord.pid
-minfds=1024
-nodaemon=true
-
-[rpcinterface:supervisor]
-supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
-
-[supervisorctl]
-serverurl=unix:///var/run/supervisor.sock
-
-[include]
-files = /etc/supervisor/conf.d/*.conf
