--- test/featuretests/images/base/supervisor_ssh.conf
@@ -1,10 +1,0 @@
-[program:sshd]
-user=root
-command=/usr/sbin/sshd -D
-autostart=true
-autorestart=true
-stdout_logfile=/proc/self/fd/1
-stdout_logfile_maxbytes=0
-stderr_logfile=/proc/self/fd/2
-stderr_logfile_maxbytes=0
-priority=10
