--- test/featuretests/images/mysql/mysync.yaml
@@ -1,57 +1,0 @@
-log: /dev/stderr
-loglevel: Debug
-db_timeout: 5s
-db_lost_check_timeout: 1s
-tick_interval: 2s
-healthcheck_interval: 5s
-dcs_wait_timeout: 10s
-failover: ${MYSYNC_FAILOVER:-false}
-failover_cooldown: ${MYSYNC_FAILOVER_COOLDOWN:-60m}
-failover_delay: ${MYSYNC_FAILOVER_DELAY:-0s}
-inactivation_delay: ${MYSYNC_INACTIVATION_DELAY:-5s}
-semi_sync: ${MYSYNC_SEMISYNC:-true}
-resetupfile: /tmp/mysync.resetup
-resetup_crashed_hosts: ${MYSYNC_RESETUP_CRASHED_HOSTS:-false}
-zookeeper:
-  session_timeout: 3s
-  namespace: /test
-  hosts: [ $ZK_SERVERS ]
-  auth: true
-  username: testuser
-  password: testpassword123
-  use_ssl: true
-  keyfile: /etc/zk-ssl/server.key
-  certfile: /etc/zk-ssl/server.crt
-  ca_cert: /etc/zk-ssl/ca.cert.pem
-  verify_certs: true
-mysql:
-  user: $MYSQL_ADMIN_USER
-  password: $MYSQL_ADMIN_PASSWORD
-  replication_user: repl
-  replication_password: repl_pwd
-  port: $MYSQL_PORT
-  external_replication_ssl_ca: /etc/mysql/ssl/external_CA.pem
-  pid_file: /tmp/mysqld.pid
-  error_log: /var/log/mysql/error.log
-disable_semi_sync_replication_on_maintenance: ${MYSYNC_DISABLE_REPLICATION_ON_MAINT:-false}
-rpl_semi_sync_master_wait_for_slave_count: ${MYSYNC_WAIT_FOR_SLAVE_COUNT:-1}
-critical_disk_usage: ${MYSYNC_CRITICAL_DISK_USAGE:-100}
-keep_super_writable_on_critical_disk_usage: ${MYSYNC_KEEP_SUPER_WRITABLE_ON_CRITICAL_DISK_USAGE:-false}
-test_disk_usage_file: /tmp/usedspace
-db_set_ro_force_timeout: 40s
-db_set_ro_timeout: ${MYSYNC_SET_RO_TIMEOUT:-30s}
-offline_mode_enable_lag: 10s
-offline_mode_disable_lag: 5s
-priority_choice_max_lag: ${MYSYNC_PRIORITY_CHOICE_LAG:-60s}
-disable_set_readonly_on_lost: false
-stream_from_reasonable_lag: ${MYSYNC_STREAM_FROM_REASONABLE_LAG:-5m}
-exclude_users:
-    - 'repl'
-    - 'something'
-    - 'admin'
-    - 'monitor'
-    - 'event_scheduler'
-replication_repair_cooldown: 10s
-replication_repair_aggressive_mode: ${MYSYNC_REPLICATION_REPAIR_AGGRESSIVE_MODE:-false}
-test_filesystem_readonly_file: /tmp/readonly
-replication_channel: ''
