--- test/featuretests/images/mysql/supervisor_mysql.conf
@@ -1,25 +1,0 @@
-[program:mysync]
-command=bash /var/lib/dist/mysql/start_mysync.sh
-process_name=%(program_name)s
-autostart=true
-startretries=100000
-autorestart=true
-stopsignal=TERM
-user=mysql
-priority=15
-stdout_logfile=/var/log/mysync.log
-stdout_logfile_maxbytes=0
-redirect_stderr=true
-
-[program:mysqld]
-command=bash /var/lib/dist/mysql/start_mysql.sh
-process_name=%(program_name)s
-autostart=true
-autorestart=false
-stopsignal=TERM
-user=mysql
-group=mysql
-priority=5
-stdout_logfile=/var/log/mysql/error.log
-stdout_logfile_maxbytes=0
-redirect_stderr=true
