--- test/featuretests/images/mysql_jepsen/.my.cnf
@@ -1,3 +1,0 @@
-[client]
-user=admin
-password=admin_pwd
