--- test/featuretests/images/mysql_jepsen/sh-scripts/my-resetup-wd.sh
@@ -1,11 +1,0 @@
-#!/bin/bash
-
-set -xe
-
-while true
-do
-    echo "=============="
-    date
-    flock -n /tmp/resetup.lock /usr/bin/my-resetup || true
-    sleep 10
-done
