--- test/featuretests/images/zookeeper/setup.sh
@@ -1,3 +1,0 @@
-set -e
-
-cp /var/lib/dist/zookeeper/supervisor_zookeeper.conf /etc/supervisor/conf.d
