--- test/featuretests/images/zookeeper/supervisor_zookeeper.conf
@@ -1,8 +1,0 @@
-[program:zookeeper]
-command=bash /var/lib/dist/zookeeper/start.sh
-environment=ZOO_LOG_DIR=/var/log/zookeeper,ZOO_LOG4J_PROP='INFO,ROLLINGFILE'
-process_name=%(program_name)s
-autostart=true
-autorestart=true
-stopsignal=TERM
-priority=5
