--- test/featuretests/images/zookeeper/zoo.cfg
@@ -1,50 +1,0 @@
-# The number of milliseconds of each tick
-tickTime=1000
-# The number of ticks that the initial
-# synchronization phase can take
-initLimit=20
-# The number of ticks that can pass between
-# sending a request and getting an acknowledgement
-syncLimit=10
-# the directory where the snapshot is stored.
-# do not use /tmp for storage, /tmp here is just
-# example sakes.
-dataDir=/tmp/zookeeper
-# the port at which the clients will connect
-clientPort=2181
-# the maximum session timeout in milliseconds that the server
-# will allow the client to negotiate.
-maxSessionTimeout=60000
-snapCount=1000000
-forceSync=no
-# the timeout value for opening connections for leader election notifications.
-cnxTimeout=3000
-# the maximum number of client connections.
-# increase this if you need to handle more clients
-#maxClientCnxns=60
-#
-# Be sure to read the maintenance section of the
-# administrator guide before turning on autopurge.
-#
-# http://zookeeper.apache.org/doc/current/zookeeperAdmin.html#sc_maintenance
-#
-# The number of snapshots to retain in dataDir
-autopurge.snapRetainCount=3
-# Purge task interval in hours
-# Set to "0" to disable auto purge feature
-autopurge.purgeInterval=0
-leaderServes=yes
-quorumListenOnAllIPs=true
-jute.maxbuffer=16777216
-secureClientPort=2281
-serverCnxnFactory=org.apache.zookeeper.server.NettyServerCnxnFactory
-portUnification=false
-skipACL=no
-ssl.trustStore.password=testpassword123
-ssl.trustStore.location=/etc/zk-ssl/truststore.jks
-ssl.keyStore.password=testpassword321
-ssl.keyStore.location=/etc/zk-ssl/server.jks
-
-# servers section will be generated from env variablses
-
-
